Refactor: 리팩토링 변경사항 반영
