+++ app/_components/header/index.tsx
@@ -8,8 +8,8 @@ import type { Response } from '@/_types/authentication';
 
 import useUserStore from '@/_stores/useUserStore';
 
-import GuestHeader from './_component/guest-header';
-import UserHeader from './_component/user-header';
+import GuestHeader from './_component/GuestHeader';
+import UserHeader from './_component/UserHeader';
 
 import logoWithTitle from 'public/assets/icons/logo-with-title.svg';
 
@@ -25,9 +25,9 @@ export default function Header() {
   }, [setLoginStatus]);
 
   return (
-    <header className="fixed left-0 right-0 top-0 z-30 w-full bg-white shadow-sm">
+    <header className="fixed left-0 right-0 top-0 md:px-[24px] z-30 w-full bg-white shadow-sm">
       <section className="mx-auto flex h-[70px] items-center justify-between px-6 py-[21px] lg:max-w-[1200px] lg:px-0">
-        <Link href="/main" className="relative h-[28px] w-auto">
+        <Link href="/" className="relative h-[28px] w-auto">
           <Image src={logoWithTitle} alt="Logo" />
         </Link>
         <nav className="">{isLoggedIn ? <UserHeader /> : <GuestHeader />}</nav>
